// src/modules/user/dtos/generated.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
  "/profile": {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** Get authenticated user profile */
    get: operations["getUserProfile"];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
}
export type webhooks = Record<string, never>;
export interface components {
  headers: never;
  parameters: never;
  pathItems: never;
  requestBodies: never;
  responses: never;
  schemas: {
    User: {
      userId: string;
      /** Format: email */
      email: string;
      isEmailVerified?: boolean;
      firstName?: string;
      lastName?: string;
      displayName?: string;
      /** Format: uri */
      avatarUrl?: string;
      /** @enum {string} */
      role?: "user" | "admin";
      isActive?: boolean;
      /** Format: date-time */
      createdAt?: string;
      /** Format: date-time */
      updatedAt?: string;
    };
    ProfileDetails: {
      firstName: string;
      lastName: string;
      bio?: string;
      phoneNumber?: string;
      countryName?: string;
      /** Format: email */
      email: string;
      /** Format: date */
      dob?: string;
      bookingsCount: number;
      favouritesCount: number;
    };
  };
}
export type $defs = Record<string, never>;
export interface operations {
  getUserProfile: {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description User profile */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          "application/json": components["schemas"]["ProfileDetails"];
        };
      };
      /** @description Unauthorized */
      401: {
        headers: {
          [name: string]: unknown;
        };
        content?: never;
      };
    };
  };
}